jtag: remove warnings
smemio.c: In function ‘GetJtagPort’:
smemio.c:40:64: warning: cast from pointer to integer of different size
[-Wpointer-to-int-cast]
   printf("GetJtagPort: find_controller: ERROR: JTAG Addr:%x\n",(int)
vmeAddress);
                                                                ^
jtag.c: In function ‘main’:
jtag.c:51:4: warning: cast from pointer to integer of different size
[-Wpointer-to-int-cast]
    (int) vmeAddress,
    ^

+++ jtag/jtag.c
@@ -48,7 +48,7 @@ int cc;
 
    printf("VHDL-Compiled BitStream Filename: %s VMEAddress: 0x%X DebugLevel: %d \n",
 	  fname,
-	  (int) vmeAddress,
+	  vmeAddress,
 	  (int) xsvf_iDebugLevel);
 
    printf("Continue (Y/N):"); yn = getchar();
